LGTM絞りフォームをコンポーネント化
